Add warning to Queue.get() docs about block=True
When Queue.get() is called with `block=True` and `timeout=None`,
the thread will block and wait on the Queue's not_empty condition
variable without a timeout. Under these conditions, per the discussion
in https://bugs.python.org/issue1360 , if the main thread is waiting on
the get(), it will not respond to SIGINT.

This can be a source of confusion in scripts which want to use Queues to
wait on async processes, as Ctrl+C is commonplace.
As noted in issue1360, this is unlikely to change because it would slow down
lock acquisition, so update the docs to include a warning instead.

I don't think there are other particularly interesting scenarios which
deserve explicit mention in the docs.
Needs to be backported to 2.7, and all 3.x maintenance branches.