Warning on Queue.get() notes change under 3.x
As of 3.0, the uninterruptible nature of Queue.get() appears to have
changed. However, the problem remains on all versions for Windows. There
doesn't seem to be prescriptive guidance for how to handle these kinds
of version-specific warnings in any of the contributing guides. Simply
describing the version/platform constraints seems fine.