bpo-16968: Clean up test_concurrent_futures
The old test_main function would cause any dangling threads or processes
to be reaped silently rather than allowing regrtest to detect and warn
about them.  Since the shutdown methods of the module should do all of
the cleanup for us, we want to get those warnings if they don't.